backend/requirements.txt
aenum==3.1.16
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==16.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
ctranslate2==4.6.2
dataclasses-json==0.6.7
decorator==5.2.1
deepgram-sdk==3.3.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.128.0
faster-whisper==1.2.1
filelock==3.20.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.4
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.41.1
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.58.1
numpy==1.26.4
onnxruntime==1.23.2
openai==2.15.0
packaging==25.0
platformdirs==4.5.1
pooch==1.9.0
propcache==0.4.1
protobuf==6.33.2
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pypdf==6.6.0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soxr==1.0.0
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
verboselogs==1.7
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
